requirements.txt
pytorch-lightning
datasets
transformers
scikit-learn
tensorboard
wandb
torchmetrics
matplotlib
seaborn